# Swagger/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Track API",
    "description" : "API of Tracks",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/JonasT3000/Band/1",
    "description" : "SwaggerHub API Auto Mocking",
    "variables" : { }
  } ],
  "paths" : {
    "/Tracks" : {
      "get" : {
        "tags" : [ "Tracks" ],
        "summary" : "Liste aller Tracks",
        "operationId" : "getTracks",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Track"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Tracks" ],
        "summary" : "Legt einen Track an",
        "operationId" : "createTrack",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TrackItem"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Created!",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                },
                "example" : "Created track: Rammstein, Engel"
              }
            }
          }
        }
      }
    },
    "/Tracks/{trackid}" : {
      "put" : {
        "tags" : [ "Tracks" ],
        "summary" : "Ändert einen Track",
        "operationId" : "changeTrack",
        "parameters" : [ {
          "name" : "trackid",
          "in" : "path",
          "description" : "ID des Tracks",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TrackItem"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "TrackItem" : {
        "type" : "object",
        "properties" : {
          "band" : {
            "maximum" : 20,
            "type" : "string",
            "example" : "Rammstein"
          },
          "title" : {
            "maximum" : 20,
            "type" : "string",
            "example" : "Engel"
          },
          "time" : {
            "maximum" : 6,
            "minimum" : 6,
            "type" : "string",
            "example" : "04m26s"
          },
          "genre" : {
            "maximum" : 20,
            "type" : "string",
            "example" : "Metal"
          }
        },
        "description" : "More infos about a Track"
      },
      "Track" : {
        "type" : "object",
        "properties" : {
          "band" : {
            "maximum" : 20,
            "type" : "string",
            "example" : "Metallica"
          },
          "title" : {
            "maximum" : 20,
            "type" : "string",
            "example" : "Master of Puppets"
          }
        },
        "description" : "Track with Band and Title"
      }
    }
  }
}